src/requirements.txt
Adafruit-MotorHAT==1.4.0
RPi.GPIO==0.7.0